refactor(modal): move useModal hook and update imports
