+++ front/src/ui/organisms/Modal/Modal.mdx
@@ -10,7 +10,7 @@ A reusable dialog component built on top of the native HTML `<dialog>` element f
 ## Usage
 
 ```tsx
-import { useModal } from '@Front/ui/utils/hooks/useModal';
+import { useModal } from '@Front/ui/utils/hooks/useModal/useModal';
 import { Modal } from '@Front/ui/organisms/modal/Modal';
 
 const Example = () => {
@@ -143,7 +143,7 @@ The component is designed to be controlled with a dialog ref (for example via `u
 
 ```tsx
 import { createPortal } from 'react-dom';
-import { useModal } from '@Front/ui/utils/hooks/useModal';
+import { useModal } from '@Front/ui/utils/hooks/useModal/useModal';
 import { Modal } from '@Front/ui/organisms/modal/Modal';
 
 const Example = () => {
